requirements.txt
aws-sam-cli==1.117.0
boto3==1.34.114
pytest==8.2.2
moto==5.0.9
python-dotenv==1.0.1